Fix emscan-deps to handle emscripten args
In order to make this work I extracted the command line handling
code into tools/cmdline.py.   This should also help for creating emld,
the standalone linker.

Without this change the test fails with:

```
Error while scanning dependencies for /usr/local/google/home/sbc/dev/wasm/emscripten/test/cmake/cxx20/main.cpp:
error: unknown argument: '-sUSE_SDL'
```

Fixes: #24454, #24410
