+++ test/cmake/cxx20/CMakeLists.txt
@@ -4,4 +4,6 @@ project(cxx20test)
 
 set(CMAKE_CXX_STANDARD 20)
 
+set(CMAKE_CXX_FLAGS "${CMAKE_CXX_FLAGS} -sUSE_SDL")
+
 add_executable(cxx20test main.cpp)
